==> step2.ipynb <==
import pandas as pd
import requests
import time

### Create Nodes for Mapping

raw = pd.read_csv('./data/mergedData/reducedStops.csv')
df = raw.drop(['Unnamed: 0', 'id'], axis=1)

# fucntion used to get summary statistics and concate station types
def f(x):
     return pd.Series(dict(lat = x['lat'].mean(),
                           long = x['long'].mean(),
                           transtype = "[%s]" % ', '.join(x['type'].unique())))

df = df.groupby('assigned').apply(f)

def concatenatorator(x):
    x = x[1:-1]
    x = x.split(',')
    x = [y.strip() for y in x]
    x.sort()
    tmp = str()
    for i in x:
        if i == "bikeStop":
            tmp = tmp+'bi'
        elif i == "busStop":
            tmp = tmp+'bu'
        elif i == "trainStop":
            tmp = tmp+'tr'
    return tmp

finaldf = df.copy()
finaldf['assigned'] = df.index.values
finaldf['convertTrans'] = finaldf['transtype'].apply(concatenatorator)
finaldf['mikeStationId'] = finaldf['convertTrans']+finaldf['assigned'].astype(str)
finaldf.drop(['transtype', 'convertTrans'], axis=1, inplace=True)

# output the nodes created from merged function to be used in Step #3
finaldf.to_csv("./data/mergedData/nodes.csv", index=False)

### Create Edges for Mapping

def concatenatorator2(x):
    tmp = str()
    if x == "bikeStop":
        tmp = 'bi'
    elif x == "busStop":
        tmp = 'bu'
    elif x == "trainStop":
        tmp = 'tr'
    return tmp

# use the raw loaded data to create a complete list of all
# mta station ids and mike's assigned ids
# this is different then the groupby approach above.
mergetable = raw.drop(['Unnamed: 0', 'lat', 'long'], axis=1)
mergetable['type2'] = mergetable['type'].apply(concatenatorator2)
mergetable['type'] = mergetable['type2']
mergetable['mtaTypeStatId'] = mergetable['type']+mergetable['id'].astype(str)
mergetable.drop('type2', axis=1, inplace=True)
mergetable.head(1)

trips = pd.read_csv('./data/mergedData/busTrainTrips.csv')


joined_dfs = mergetable.merge(finaldf, how='inner', on='assigned')
joined_dfs.head(1)

#merge old and new id's in trips

joined_dfs['id'] =joined_dfs['id'].apply(lambda x: str(x).split('.')[0])
trips['mta_start_id'] = trips['mta_start_id'].apply(lambda x: str(x).split('.')[0])
trips['mta_stop_id'] = trips['mta_stop_id'].apply(lambda x: str(x).split('.')[0])


keyValuePair= joined_dfs[['id','mikeStationId']]
keyValuePair = keyValuePair.drop_duplicates()
ids = list(keyValuePair['id'].values)
reducedId = list(keyValuePair['mikeStationId'].values)

switchingDic = dict(zip(ids,reducedId))

trips['source']=trips['mta_start_id'].apply(lambda x: switchingDic[str(x)])
trips['target']=trips['mta_stop_id'].apply(lambda x: switchingDic[str(x)])


tripsReduced = trips[['source','target','duration']]
tripsReduced.to_csv('./data/mergedData/reducedTrips.csv')

#reduce trips 
tripsReduced2 = tripsReduced.groupby(['source','target']).mean()
tripsReduced2= tripsReduced2.reset_index()
tripsReduced2.to_csv('./data/mergedData/averagedTripDuration.csv')
